fix: remove config.go and inline it in syncer.go
