+++ plugin/evm/atomic/sync/summary_test.go
@@ -17,9 +17,7 @@ import (
 
 func TestMarshalSummary(t *testing.T) {
 	atomicSummary, err := NewSummary(common.Hash{1}, 2, common.Hash{3}, common.Hash{4})
-	if err != nil {
-		t.Fatalf("failed to create summary: %v", err)
-	}
+	require.NoError(t, err, "failed to create summary")
 
 	require.Equal(t, common.Hash{1}, atomicSummary.GetBlockHash())
 	require.Equal(t, uint64(2), atomicSummary.Height())
@@ -37,9 +35,7 @@ func TestMarshalSummary(t *testing.T) {
 		return block.StateSyncSkipped, nil
 	}
 	s, err := parser.Parse(atomicSummary.Bytes(), acceptImplTest)
-	if err != nil {
-		t.Fatalf("failed to parse summary: %v", err)
-	}
+	require.NoError(t, err, "failed to parse summary")
 	require.Equal(t, atomicSummary.GetBlockHash(), s.GetBlockHash())
 	require.Equal(t, atomicSummary.Height(), s.Height())
 	require.Equal(t, atomicSummary.GetBlockRoot(), s.GetBlockRoot())
@@ -47,9 +43,7 @@ func TestMarshalSummary(t *testing.T) {
 	require.Equal(t, atomicSummary.Bytes(), s.Bytes())
 
 	mode, err := s.Accept(context.TODO())
-	if err != nil {
-		t.Fatalf("failed to accept summary: %v", err)
-	}
+	require.NoError(t, err, "failed to accept summary")
 	require.Equal(t, block.StateSyncSkipped, mode)
 	require.True(t, called)
 }
